+++ src/CMakeLists.txt
@@ -25,8 +25,6 @@ target_include_directories(${ASM_2_BIN} PUBLIC ../include)
 target_include_directories(${ASM_2_BIN} PUBLIC ${CMAKE_CURRENT_BINARY_DIR})
 
 add_custom_target(generate_all)
-add_dependencies(${LIB_NAME} generate_all)
-add_dependencies(${ASM_2_BIN} generate_all)
 
 set(TEMPLATES 
     ${CMAKE_CURRENT_SOURCE_DIR}/templates/Instruction_enum.h.temp
@@ -48,3 +46,6 @@ foreach(TEMPL ${TEMPLATES})
     add_custom_target(${TARGET_NAME}_gen DEPENDS ${CMAKE_CURRENT_BINARY_DIR}/generated/${OUTPUT_NAME_TEMPL})
     add_dependencies(generate_all ${TARGET_NAME}_gen)
 endforeach()
+
+add_dependencies(${LIB_NAME} generate_all)
+add_dependencies(${ASM_2_BIN} Decoder_gen Instruction_enum_gen)
