+++ src/templates/Decoder.cpp.temp
@@ -1,5 +1,5 @@
 #include "Decoder.h"
-
+#include <iostream>
 namespace {
 
 /* [[[cog
