+++ tests/CMakeLists.txt
@@ -1,13 +1,7 @@
 add_custom_target(run_all_tests)
 
 set(TESTS
-    square_eq
-    arithmetic
-    branching
-    intrinsic
-    call
-    array
-    string
+    tests
 )
 
 foreach(EXEC IN LISTS TESTS)
